Add ability to submit an ad via browser form
It's rudementry, and there is no authentication required to post. That
is, you don't need to be a registered user to post.

Things are still messy. The fields have no labels; their coming soon. No
ability to edit or delete either. The latter will definetly require
being a logged-in user.

Remember, we're only in testing days at present. Patience.